Tweaked imports to avoid loading Monaco editor unless needed
